Moved `load_secrets()` and `save_secrets` to config.py. (#1)
