+++ mipcandy/__init__.py
@@ -1,4 +1,5 @@
 from mipcandy.common import *
+from mipcandy.config import load_secrets, save_secrets
 from mipcandy.data import *
 from mipcandy.evaluation import EvalCase, EvalResult, Evaluator
 from mipcandy.frontend import *
