+++ mipcandy/frontend/__init__.py
@@ -1,2 +1,2 @@
 from mipcandy.frontend.notion_fe import NotionFrontend
-from mipcandy.frontend.prototype import load_secrets, save_secrets, Frontend, create_hybrid_frontend
+from mipcandy.frontend.prototype import Frontend, create_hybrid_frontend
